refactor: 지출/수입 내역 디렉토리 구조 변경
