+++ src/client/modules/Account/AccountHistory/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './AccountHistory';
